requirements.txt
astroid==2.3.1
certifi==2019.9.11
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
Pillow==6.1.0
pylint==2.4.2
python-pptx==0.6.18
six==1.12.0
wrapt==1.11.2
XlsxWriter==1.2.1
